Move karumo/shot screenshot tests from the sample app to the ui-android library module.
