+++ android/ui-android/build.gradle.kts
@@ -1,5 +1,20 @@
 plugins {
     id("android-library-convention")
+    id("shot")
+}
+
+android {
+    defaultConfig {
+        testApplicationId = "com.trafi.ui.test"
+        testInstrumentationRunner = "com.karumi.shot.ShotTestRunner"
+    }
+    // Workaround for https://github.com/Kotlin/kotlinx.coroutines/issues/2023
+    packagingOptions {
+        resources {
+            excludes += "META-INF/AL2.0"
+            excludes += "META-INF/LGPL2.1"
+        }
+    }
 }
 
 dependencies {
@@ -10,4 +25,6 @@ dependencies {
     implementation("androidx.compose.runtime:runtime:${Versions.compose}")
     implementation("androidx.compose.material:material:${Versions.compose}")
     implementation("androidx.compose.ui:ui-tooling:${Versions.compose}")
+
+    androidTestImplementation("androidx.test:rules:1.3.0")
 }
