+++ android/ui-android/src/androidTest/java/com/trafi/ui/PreviewTest.kt
@@ -1,10 +1,10 @@
-package com.trafi.example
+package com.trafi.ui
 
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.test.junit4.ComposeContentTestRule
 import com.karumi.shot.ScreenshotTest
 
-class PreviewTest(private val composeTestRule: ComposeContentTestRule) : ScreenshotTest {
+internal class PreviewTest(private val composeTestRule: ComposeContentTestRule) : ScreenshotTest {
     fun renderAndTestView(screenshotName: String, composable: @Composable () -> Unit) {
         composeTestRule.setContent(composable)
         compareScreenshot(composeTestRule, screenshotName)
