+++ android/ui-android/src/androidTest/java/com/trafi/ui/PreviewTests.kt
@@ -1,9 +1,10 @@
-package com.trafi.example
+package com.trafi.ui
 
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.test.junit4.ComposeContentTestRule
 import androidx.compose.ui.test.junit4.createComposeRule
 import androidx.test.rule.GrantPermissionRule
+import androidx.test.rule.GrantPermissionRule.grant
 import com.karumi.shot.ScreenshotTest
 import org.junit.Before
 import org.junit.Rule
@@ -12,37 +13,34 @@ import org.junit.runner.RunWith
 import org.junit.runners.Parameterized
 
 @RunWith(Parameterized::class)
-class ComposablePreviewsTests(
+public class PreviewTests(
     private val previewWithName: Pair<@Composable () -> Unit, String>,
 ) : ScreenshotTest {
-    companion object {
+    public companion object {
         @JvmStatic
         @Parameterized.Parameters
-        fun previews() = testPreviews
+        public fun previews(): List<Pair<@Composable () -> Unit, String>> = testPreviews
     }
 
-    lateinit var previewTest: PreviewTest
+    private lateinit var previewTest: PreviewTest
 
     @get:Rule
-    val composeTestRule: ComposeContentTestRule = createComposeRule()
+    public val composeTestRule: ComposeContentTestRule = createComposeRule()
 
     @get:Rule
-    val grantPermissionRule = GrantPermissionRule.grant(
+    public val grantPermissionRule: GrantPermissionRule = grant(
         android.Manifest.permission.READ_EXTERNAL_STORAGE,
         android.Manifest.permission.WRITE_EXTERNAL_STORAGE,
     )
 
     @Before
-    fun init() {
+    public fun init() {
         previewTest = PreviewTest(composeTestRule)
     }
 
     @Test
-    fun previewTest() {
+    public fun previewTest() {
         val (preview, name) = previewWithName
-        previewTest
-            .renderAndTestView(name) {
-                preview()
-            }
+        previewTest.renderAndTestView(name) { preview() }
     }
 }
